+++ web/src/components/App.tsx
@@ -14,7 +14,7 @@ import { composeWithDevTools } from '@redux-devtools/extension'
 import { createBrowserHistory } from 'history'
 import { theme } from '../helpers/theme'
 import BottomBar from './bottom-bar/BottomBar'
-import ColumnLevel from './column-level/ColumnLevel'
+import ColumnLevel from '../routes/column-level/ColumnLevel'
 import Datasets from '../routes/datasets/Datasets'
 import Events from '../routes/events/Events'
 import Header from './header/Header'
