+++ web/src/routes/column-level/ZoomControls.tsx
